gg
gg

+++ src/main.cpp
@@ -275,7 +275,7 @@ void boot_screen_anim() {
     }
     tft.drawPixel(0, 0, 0);       // Forces back communication with TFT, to avoid ghosting
     // Start image loop
-    while (millis() < i + 7000) { // boot image lasts for 5 secs
+    while (millis() < (unsigned long)(i + 7000)) { // boot image lasts for 5 secs
         if ((millis() - i > 2000) && !drawn) {
             tft.fillRect(0, 45, tftWidth, tftHeight - 45, bruceConfig.bgColor);
             if (boot_img > 0 && !drawn) {
